requirement.txt
# Core Dependencies
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.3.1
transformers==4.36.2

# Vector Databases
faiss-cpu==1.7.4
chromadb==0.4.22

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
tiktoken==0.5.2

# GUI
# Tkinter comes built-in with Python

# Utilities
numpy==1.24.3
pandas==2.0.3
tqdm==4.66.1
